Restructured prompt
Updated weapon_info to include a trailing 'and' for better sentence structure regardless of if the character is carrying a weapon or not. Faction info delivered as a "who is" statement about the speaking character in the beginning of the prompt. Personality delivered as "who is [blank]" instead of "in a [blank] manner" which seems to give better results and allow for more freedom, especially when creating descriptions for unique characters